reference/mdb-bfsi-credit-reco-genai/backend_agentic/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
joblib
pymongo
certifi
python-dotenv
scipy
xgboost
torch
typing-extensions
huggingface_hub==0.16.4
sentence-transformers==2.2.2
InstructorEmbedding
langchain==0.3.24
langchain-community
langchain-core
langchain-fireworks
langchain-mongodb
langchain-text-splitters==0.3.8
langgraph==0.4.1
langgraph-checkpoint-mongodb==0.1.3


